+++ src/mnemonic.cpp
@@ -3,11 +3,10 @@
 #include "mnemonic_bit.hpp"
 #include "mnemonic_branch.hpp"
 #include "mnemonic_calc.hpp"
-#include "mnemonic_dec.hpp"
 #include "mnemonic_defb.hpp"
 #include "mnemonic_defw.hpp"
 #include "mnemonic_ex.hpp"
-#include "mnemonic_inc.hpp"
+#include "mnemonic_increment.hpp"
 #include "mnemonic_io.hpp"
 #include "mnemonic_load.hpp"
 #include "mnemonic_shift.hpp"
